fix(events): Separate the events context
