--- packages/module/src/DataViewContext/index.ts
@@ -1,2 +1,0 @@
-export { default } from './DataViewContext';
-export * from './DataViewContext';
