+++ packages/module/src/DataViewEventsContext/index.ts
@@ -1,0 +1,2 @@
+export { default } from './DataViewEventsContext';
+export * from './DataViewEventsContext';
